cache stdin data in package var and moved stringSliceToMap to utils.go
