+++ data_reader.go
@@ -1,10 +1,7 @@
 package main
 
 import (
-	"errors"
-	"fmt"
 	"io/ioutil"
-	"os"
 )
 
 type sourceReader interface {
@@ -30,17 +27,18 @@ func readData(inputStrSlice []string, sr sourceReader) (map[string]interface{}, 
 
 		byteData := []byte{}
 		if sr.IsFile(value) {
+			// data from file
 			byteData, err = ioutil.ReadFile(value)
 			if err != nil {
 				return nil, err
 			}
 		} else if value == "-" {
 			// data from stdin
-			stdinData, err := ioutil.ReadAll(os.Stdin)
+			stdin, err := readStdinData()
 			if err != nil {
-				return nil, fmt.Errorf("unable read stdin data: %w", err)
+				return nil, err
 			}
-			byteData = stdinData
+			byteData = stdin
 		} else {
 			// inline data
 			byteData = []byte(value)
@@ -54,20 +52,4 @@ func readData(inputStrSlice []string, sr sourceReader) (map[string]interface{}, 
 		retData[key] = data
 	}
 	return retData, nil
-}
-
-// stringSliceToMap extracts key=value pairs from the string slice and writes them as key=value pair to the map
-func stringSliceToMap(strs []string) (map[string]string, error) {
-	if len(strs) < 1 {
-		return nil, errors.New("need at least one key=value pair")
-	}
-	rslt := map[string]string{}
-	for _, idxValue := range strs {
-		key, value, err := splitVarParam(idxValue)
-		if err != nil {
-			return nil, err
-		}
-		rslt[key] = value
-	}
-	return rslt, nil
 }
