+++ data_reader_test.go
@@ -154,74 +154,3 @@ func Test_readData(t *testing.T) {
 		})
 	}
 }
-
-func Test_stringSliceToMap(t *testing.T) {
-	type args struct {
-		strs []string
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    map[string]string
-		wantErr bool
-	}{
-		{
-			name: "simple key=value pair",
-			args: args{
-				strs: []string{"key=value"},
-			},
-			want: map[string]string{
-				"key": "value",
-			},
-			wantErr: false,
-		},
-		{
-			name: "long foooooooo=baaaaaaaar pair",
-			args: args{
-				strs: []string{"foooooooo=baaaaaaaar"},
-			},
-			want: map[string]string{
-				"foooooooo": "baaaaaaaar",
-			},
-			wantErr: false,
-		},
-		{
-			name: "complex key123=value123 pair",
-			args: args{
-				strs: []string{"key123=value123"},
-			},
-			want: map[string]string{
-				"key123": "value123",
-			},
-			wantErr: false,
-		},
-		{
-			name: "failure key-value pair",
-			args: args{
-				strs: []string{"key-value"},
-			},
-			want:    nil,
-			wantErr: true,
-		},
-		{
-			name: "nil value pair",
-			args: args{
-				strs: nil,
-			},
-			want:    nil,
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := stringSliceToMap(tt.args.strs)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("stringSliceToMap() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("stringSliceToMap() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
